(#57) Introduce sphere module
